handle reference bundle better
- rename jenkinshelpers->cwrhelpers
- lookup app name for a given charm in a given bundle
- use app name when deploying the local charm with a reference bundle
- include theblues to handle charmstore interaction (for retrieving bundle.yaml)
