+++ wheelhouse.txt
@@ -1,2 +1,3 @@
 jujubigdata>=7.0.0,<8.0.0
 python-jenkins>=0.4.13,<1.0.0
+theblues
